requirements.txt
alembic==1.6.5
asgiref==3.4.0
Babel==2.9.1
bcrypt==3.2.0
bidict==0.21.2
blinker==1.4
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Django==3.2.4
django-azure-sql-backend==2.1.1.1
djangorestframework==3.12.4
dnspython==1.16.0
email-validator==1.1.3
eventlet==0.31.0
greenlet==1.1.0
httpagentparser==1.9.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
ldap3==2.9
Mako==1.1.4
MarkupSafe==2.0.1
msal==1.12.0
paramiko==2.7.2
passlib==1.7.4
psutil==5.8.0
psycopg2==2.9.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
pyodbc==4.0.30
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
pywinpty==1.1.1
requests==2.25.1
simplejson==3.17.2
six==1.16.0
speaklater==1.3
speaklater3==1.4
sqlparse==0.4.1
sshtunnel==0.4.0
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
Werkzeug==1.0.1
WTForms==2.3.3